requirements.txt
click==6.6
flask==1.1.1
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==0.23
pyaml==16.9.0
pyyaml>=4.2b1
requests>=2.20.0
six==1.10.0
slackclient==1.0.2
websocket-client==0.37.0
Werkzeug==0.15
sumy==0.7.0
numpy==1.16.2
google-cloud-language==1.2.0
nltk==3.4
